+++ src/main/java/zimareva/repository/ItemRepository.java
@@ -6,11 +6,12 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import zimareva.model.Item;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
 public interface ItemRepository extends CrudRepository<Item, Long> {
     @Query(nativeQuery = true)
-    List<Long> findIdsByColorAndBoxId(@Param("idBox") Long idBox,
-                                      @Param("color") String color);
+    List<BigInteger> findIdsByColorAndBoxId(@Param("idBox") Long idBox,
+                                            @Param("color") String color);
 }
